Move win_registry module to parsec.core
