+++ parsec/core/win_registry.py
@@ -1,0 +1,86 @@
+# Parsec Cloud (https://parsec.cloud) Copyright (c) AGPLv3 2019 Scille SAS
+
+import platform
+from structlog import get_logger
+
+logger = get_logger()
+
+ACROBAT_READER_DC_PRIVILEGED = "Software\\Adobe\\Acrobat Reader\\DC\\Privileged"
+ENABLE_APP_CONTAINER = "bEnableProtectedModeAppContainer"
+
+
+# Acrobat container app workaround
+
+
+def winreg_read_user_dword(key, name):
+    import winreg
+
+    hkcu = winreg.HKEY_CURRENT_USER
+    with winreg.OpenKey(hkcu, key, access=winreg.KEY_READ) as key_obj:
+        value, value_type = winreg.QueryValueEx(key_obj, name)
+    assert value_type == winreg.REG_DWORD
+    return value
+
+
+def winreg_write_user_dword(key, name, value):
+    import winreg
+
+    hkcu = winreg.HKEY_CURRENT_USER
+    with winreg.OpenKey(hkcu, key, access=winreg.KEY_SET_VALUE) as key_obj:
+        winreg.SetValueEx(key_obj, name, None, winreg.REG_DWORD, value)
+
+
+def winreg_delete_user_dword(key, name):
+    import winreg
+
+    hkcu = winreg.HKEY_CURRENT_USER
+    with winreg.OpenKey(hkcu, key, access=winreg.KEY_SET_VALUE) as key_obj:
+        winreg.DeleteValue(key_obj, name)
+
+
+def winreg_has_user_key(key):
+    import winreg
+
+    hkcu = winreg.HKEY_CURRENT_USER
+    try:
+        with winreg.OpenKey(hkcu, key, access=winreg.KEY_READ):
+            return True
+    except FileNotFoundError:
+        return False
+
+
+def is_acrobat_reader_dc_present():
+    if platform.system() != "Windows":
+        return False
+
+    try:
+        import winreg  # noqa
+    except ImportError:
+        logger.warning("OS is Windows but cannot import winreg")
+        return False
+
+    return winreg_has_user_key(ACROBAT_READER_DC_PRIVILEGED)
+
+
+def get_acrobat_app_container_enabled():
+    if not is_acrobat_reader_dc_present():
+        return False
+    try:
+        value = winreg_read_user_dword(ACROBAT_READER_DC_PRIVILEGED, ENABLE_APP_CONTAINER)
+    except FileNotFoundError:
+        # If the value doesn't exist, Acrobat considers it true
+        return True
+
+    return bool(value)
+
+
+def set_acrobat_app_container_enabled(value):
+    if not is_acrobat_reader_dc_present():
+        return
+    winreg_write_user_dword(ACROBAT_READER_DC_PRIVILEGED, ENABLE_APP_CONTAINER, value)
+
+
+def del_acrobat_app_container_enabled():
+    if not is_acrobat_reader_dc_present():
+        return
+    winreg_delete_user_dword(ACROBAT_READER_DC_PRIVILEGED, ENABLE_APP_CONTAINER)
